+++ src/tools/trackgen/util.cpp
@@ -38,6 +38,7 @@
 #include <tgfclient.h>
 #include <track.h>
 #include <robottools.h>
+#include <portability.h>
 #include "trackgen.h"
 #include "util.h"
 
@@ -411,7 +412,7 @@ std::ostream &operator << (std::ostream &out, tSegType type)
     out << (type == TR_RGT ? "TR_RGT" :
             type == TR_LFT ? "TR_LFT" :
             type == TR_STR ? "TR_STR" :
-            std::to_string(static_cast<unsigned long long>(type)));
+            std::to_string(type));
 
     return out;
 }
@@ -423,7 +424,7 @@ std::ostream &operator << (std::ostream &out, tSegType2 type2)
             type2 == TR_RSIDE ? "TR_RSIDE" :
             type2 == TR_LBORDER ? "TR_LBORDER" :
             type2 == TR_RBORDER ? "TR_RBORDER" :
-            std::to_string(static_cast<unsigned long long>(type2)));
+            std::to_string(type2));
 
     return out;
 }
@@ -435,7 +436,7 @@ std::ostream &operator << (std::ostream &out, tSegStyle style)
             style == TR_WALL ? "TR_WALL" :
             style == TR_FENCE ? "TR_FENCE" :
             style == TR_PITBUILDING ? "TR_PITBUILDING" :
-            std::to_string(static_cast<unsigned long long>(style)));
+            std::to_string(style));
 
     return out;
 }
@@ -491,7 +492,7 @@ void dumpSeg(std::ofstream &ofs,const tTrackSeg *seg, const std::string & indent
     ofs << indent << "vertex[2]     " << seg->vertex[2] << std::endl;
     ofs << indent << "vertex[3]     " << seg->vertex[3] << std::endl;
     ofs << indent << "angle         ";
-    for (int i = 0; i < 7; ++i)
+    for (size_t i = 0; i < 7; ++i)
         ofs << seg->angle[i] << " ";
     ofs << std::endl;
     ofs << indent << "sin           " << seg->sin << std::endl;
